+++ FSIC/classes/__init__.py
@@ -5,9 +5,9 @@ classes
 FSIC subpackage of class definitions.
 
 Defines:
- - Schematic classes: Equation, Function, Model
+ - Schematic classes: Variable, Equation, Function, Model
 
 """
 
 
-from FSIC.classes.schematic import Equation, Function, Model
+from FSIC.classes.schematic import Variable, Equation, Function, Model