+++ FSIC/classes/tests/test_equation.py
@@ -10,18 +10,24 @@ def test_init_parse_default():
     eqn = Equation('C_d = alpha_1 * YD + alpha_2 * H_h[-1]')
     assert eqn.n == ['C_d']
     assert eqn.x == ['alpha_1', 'YD', 'alpha_2', 'H_h[-1]']
+    assert eqn.expr == '%s = %s * %s + %s * %s'
+    assert eqn.count == 5
 
 def test_init_parse_sep():
     eqn = Equation('C_d : alpha_1, YD, alpha_2, H_h[-1]',
                    sep=':')
     assert eqn.n == ['C_d']
     assert eqn.x == ['alpha_1', 'YD', 'alpha_2', 'H_h[-1]']
+    assert eqn.expr == '%s : %s, %s, %s, %s'
+    assert eqn.count == 5
 
 def test_init_parse_regex():
     eqn = Equation('C_d = alpha_1 * YD + alpha_2 * H_h[-1]',
                    regex=re.compile(r'\b[A-Z][_0-z]+\b'))
     assert eqn.n == ['C_d']
     assert eqn.x == ['YD', 'H_h']
+    assert eqn.expr == '%s = alpha_1 * %s + alpha_2 * %s[-1]'
+    assert eqn.count == 3
 
 def test_init_parse_both():
     eqn = Equation('C_d : alpha_1, YD, alpha_2, H_h[-1]',
@@ -29,6 +35,8 @@ def test_init_parse_both():
                    regex=re.compile(r'\b[A-Z][_0-z]+\b'))
     assert eqn.n == ['C_d']
     assert eqn.x == ['YD', 'H_h']
+    assert eqn.expr == '%s : alpha_1, %s, alpha_2, %s[-1]'
+    assert eqn.count == 3
 
 
 def test_parse_string_arg_default():
@@ -36,6 +44,8 @@ def test_parse_string_arg_default():
     eqn.parse('C_d = alpha_1 * YD + alpha_2 * H_h[-1]')
     assert eqn.n == ['C_d']
     assert eqn.x == ['alpha_1', 'YD', 'alpha_2', 'H_h[-1]']
+    assert eqn.expr == '%s = %s * %s + %s * %s'
+    assert eqn.count == 5
 
 def test_parse_string_arg_sep_str():
     eqn = Equation()
@@ -43,6 +53,8 @@ def test_parse_string_arg_sep_str():
               sep=':')
     assert eqn.n == ['C_d']
     assert eqn.x == ['alpha_1', 'YD', 'alpha_2', 'H_h[-1]']
+    assert eqn.expr == '%s : %s, %s, %s, %s'
+    assert eqn.count == 5
 
 def test_parse_string_arg_sep_re():
     eqn = Equation()
@@ -50,6 +62,8 @@ def test_parse_string_arg_sep_re():
               sep=re.compile(r':'))
     assert eqn.n == ['C_d']
     assert eqn.x == ['alpha_1', 'YD', 'alpha_2', 'H_h[-1]']
+    assert eqn.expr == '%s : %s, %s, %s, %s'
+    assert eqn.count == 5
 
 def test_parse_string_arg_regex_str():
     eqn = Equation()
@@ -57,6 +71,8 @@ def test_parse_string_arg_regex_str():
               regex=r'\b[A-Z][_0-z]+\b')
     assert eqn.n == ['C_d']
     assert eqn.x == ['YD', 'H_h']
+    assert eqn.expr == '%s = alpha_1 * %s + alpha_2 * %s[-1]'
+    assert eqn.count == 3
 
 def test_parse_string_arg_regex_re():
     eqn = Equation()
@@ -64,6 +80,8 @@ def test_parse_string_arg_regex_re():
               regex=re.compile(r'\b[A-Z][_0-z]+\b'))
     assert eqn.n == ['C_d']
     assert eqn.x == ['YD', 'H_h']
+    assert eqn.expr == '%s = alpha_1 * %s + alpha_2 * %s[-1]'
+    assert eqn.count == 3
 
 
 if __name__ == '__main__':